style(git): merge and sort ignored Python clutter
It's not as important to know where ignored clutter comes from as it is
to know ignored clutter can be located in .gitignore. To this end
ignored files from various sources are merged, sorted, and deduplicated
after commits are made showing sources' original order in an attempt to
"respect des fonds" of the recommender providing the ignore expressions.
In this way sources' orginal order can be tracked in version control
history without impedeing discovery of ignore expressions.

The exception to this is sorting by language. Expression are sorted
lexically in ascending order and are grouped by programming langauge.