refactor(git): follow contraction convention
Why:

Aliases should be based on contraction of their full name
to make the alias and the aliased command easier to remember.
In other words, aliases should reduced typing AND reinforce
learning the actual command.

This change addresses the need by:
- renames dfg (diff gui) and diff (diff cli) to dt and dtc
for difftool and difftool cli respectively. This defaults
difftool to using the configured GUI tool and shortens
access to difftool CLI.