refactor(git): differentiate stored vs sourced files
Why:

The development/git.sh file is intended for use with the basic
terminal setup package that sources all files linked into so those
scripts are available as shortcuts/configuration at the CLI.
However, adding the storage directory revealed the directories

The development/git.sh files is intended for use with terminal
dotfiles package which sources files linked into its directories
so that those scripts are available as shortcuts/configuration at
the CLI. However, when the terminal package was originally designed
it didn't account for the need for storaging files it wouldn't
source such as the .dotfiles/storage/git/precommit/python.yaml and
install.sh files. These files would break the terminal package.

This change addresses the need by:
- namespaces .dotfiles directories per package
- creates a storage directory also namespaced per package
- still allows for files to be linked to the terminal package
directory with GNU Stow